refactor avro source methods to make them available in all stream jobs
